Support prebuilding npm packages as external bundles
Implement opt-in prebundling of npm packages as external bundles. This
is opt-in because there are many packages where it may not work as
expected (eg. due to ambiguity over how to translate CommonJS exports
to ES exports in CommonJS packages) and changes are also likely to be
required in consuming projects to serve the bundles.

In the process of implementing this the existing entry point was
refactored to move most of the plugin caching logic into
`lib/plugin.js`.